Professor
upadate funcition for delete by ID and delete all for now throw erro 500

+++ src/main/java/com/project/professor/allocation/service/ProfessorService.java
@@ -6,17 +6,21 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.project.professor.allocation.entity.Professor;
+import com.project.professor.allocation.repository.AllocationRepository;
 import com.project.professor.allocation.repository.ProfessorRepository;
+import com.project.professor.allocation.service.exception.AllocationExistsException;
 import com.project.professor.allocation.service.exception.EntityNotFindException;
 
 @Service
 public class ProfessorService {
 
 	private final ProfessorRepository professorRepository;
+	private final AllocationRepository allocationRepository;
 
-	public ProfessorService(ProfessorRepository professorRepository) {
+	public ProfessorService(ProfessorRepository professorRepository, AllocationRepository allocationRepository) {
 		super();
 		this.professorRepository = professorRepository;
+		this.allocationRepository = allocationRepository;
 	}
 
 	public List<Professor> findAll() {
@@ -54,15 +58,27 @@ public class ProfessorService {
 			throw new EntityNotFindException("Professor  doesnt exists");
 	}
 
-	public void deleteById(Long id) throws EntityNotFindException {
+	public void deleteById(Long id) throws EntityNotFindException, AllocationExistsException {
 		if (id != null && professorRepository.existsById(id)) {
-			professorRepository.deleteById(id);
+			if (allocationRepository.findByProfessorId(id) == null) {
+				professorRepository.deleteById(id);
+			} else {
+				throw new AllocationExistsException("Professor have allocation");
+			}
 		} else {
 			throw new EntityNotFindException("Professor Id doesnt exists");
 		}
 	}
 
-	public void deleteAll() {
+	public void deleteAll() throws AllocationExistsException {
+
+		List<Professor> professors = findAll();
+
+		for (Professor professor : professors) {
+			if (allocationRepository.findByProfessorId(professor.getId()) != null) {
+				throw new AllocationExistsException("Professor have allocation");
+			}
+		}
 		professorRepository.deleteAll();
 	}
